feat: move getRouteId to app tools
